перенес join animal user в классы user, сделал валидацию телефона и email, отредактировал java doc
