+++ src/main/java/com/skypro/telegram_team/services/UserService.java
@@ -1,7 +1,9 @@
 package com.skypro.telegram_team.services;
 
 import com.skypro.telegram_team.exceptions.InvalidDataException;
+import com.skypro.telegram_team.models.Animal;
 import com.skypro.telegram_team.models.User;
+import com.skypro.telegram_team.repositories.AnimalRepository;
 import com.skypro.telegram_team.repositories.UserRepository;
 import lombok.extern.log4j.Log4j2;
 import org.modelmapper.ModelMapper;
@@ -14,26 +16,37 @@ import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
+import java.util.regex.Pattern;
 
 /**
- * Сервис для работы с пользователями
+ * Сервис для работы с пользователями.
+ * <p>
+ * Реализует CRUD-операции (создание, чтение, обновление, удаление) для сущности User
+ * с использованием JpaRepository.
  */
 @Log4j2
 @Service
 public class UserService {
     private final UserRepository userRepository;
+    private final AnimalRepository animalRepository;
 
-    public UserService(UserRepository userRepository) {
+    /**
+     * Конструктор класса UserService.
+     *
+     * @param userRepository   Репозиторий для работы с сущностью User.
+     * @param animalRepository Репозиторий для работы с сущностью Animal.
+     */
+    public UserService(UserRepository userRepository, AnimalRepository animalRepository) {
         this.userRepository = userRepository;
+        this.animalRepository = animalRepository;
     }
 
     /**
-     * сохранение пользователя в БД используя метод репозитория {@link JpaRepository#save(Object)}
+     * Создает нового пользователя и сохраняет его в БД.
      *
-     * @param user
-     * @return
-     * @see
+     * @param user Объект типа User с данными нового пользователя.
+     * @return Объект типа User, сохраненный в БД.
+     * @throws InvalidDataException Если данные пользователя некорректны.
      */
     @Transactional
     public User save(User user) {
@@ -79,11 +92,15 @@ public class UserService {
     }
 
     /**
-     * обновление пользователя в БД используя метод репозитория {@link JpaRepository#save(Object)}
+     * Обновляет данные пользователя в БД по заданному идентификатору, используя
+     * метод репозитория {@link JpaRepository#save(Object)}
      *
-     * @param user
-     * @param id
-     * @return
+     * @param user объект пользователя, содержащий данные для обновления
+     * @param id   идентификатор пользователя, данные которого нужно обновить
+     * @return объект пользователя после обновления данных
+     * @throws EntityNotFoundException  если пользователь с заданным идентификатором не найден в БД
+     * @throws IllegalArgumentException если объект пользователя не проходит валидацию
+     * @see ModelMapper
      */
     @Transactional
     public User update(User user, Long id) {
@@ -101,11 +118,12 @@ public class UserService {
     }
 
     /**
-     * Назначить пользователя волонтером
+     * Устанавливает флаг волонтера для пользователя.
      *
-     * @param id
-     * @param isVolunteer
-     * @return
+     * @param id          идентификатор пользователя.
+     * @param isVolunteer флаг волонтера.
+     * @return обновленный экземпляр пользователя.
+     * @throws EntityNotFoundException если пользователь не найден в БД.
      */
     @Transactional
     public User userIsVolunteer(Long id, Boolean isVolunteer) {
@@ -115,8 +133,16 @@ public class UserService {
         return update(user, id);
     }
 
+    /**
+     * Возвращает список пользователей из БД, чей статус соответствует переданному значению
+     *
+     * @param state статус владельца
+     * @return список пользователей, чей статус соответствует переданному значению
+     * @see User
+     */
     public List<User> findByState(User.OwnerStateEnum state) {
         return userRepository.findByState(state);
+    }
 
     /**
      * Поиск волонтеров
@@ -128,9 +154,10 @@ public class UserService {
     }
 
     /**
-     * Поиск любого волонтера
+     * Возвращает любого волонтера из списка волонтеров.
      *
-     * @return
+     * @return объект типа Optional<User>, содержащий любого волонтера из списка волонтеров,
+     * если список не пустой, иначе пустой объект Optional.
      */
     public Optional<User> findAnyVolunteer() {
         return findVolunteers().stream()
@@ -138,10 +165,11 @@ public class UserService {
     }
 
     /**
-     * Поиск пользователя по telegramId, возвращает пустой User если не нашел в БД
+     * Находит пользователя по заданному идентификатору Telegram.
      *
-     * @param telegramId
-     * @return
+     * @param telegramId идентификатор Telegram, по которому нужно найти пользователя.
+     * @return объект типа User, соответствующий заданному идентификатору Telegram,
+     * или пустой объект User, если пользователь не найден.
      */
     public User findByTelegramId(Long telegramId) {
         return userRepository.findByTelegramId(telegramId).stream()
@@ -150,9 +178,11 @@ public class UserService {
     }
 
     /**
-     * Проверить данные пользователя, если данные некорректны то выбросить исключение {@link InvalidDataException}
+     * Проверяет данные пользователя на корректность и выбрасывает исключение {@link InvalidDataException},
+     * если данные некорректны.
      *
-     * @param user
+     * @param user объект типа User, содержащий данные пользователя, которые нужно проверить.
+     * @throws InvalidDataException если данные пользователя некорректны.
      */
     private void validate(User user) {
         if (user.getName() == null) {
@@ -170,25 +200,43 @@ public class UserService {
     }
 
     /**
-     * Проверить телефон, вернуть true если соответствует шаблону
+     * Проверяет, соответствует ли заданный номер телефона российскому формату.
      *
-     * @param phone
-     * @return
+     * @param phone номер телефона, который нужно проверить.
+     * @return true, если номер телефона соответствует российскому формату, и false в противном случае.
      */
     private boolean isPhoneValid(String phone) {
-        //Добавить проверку телефона
-        return true;
+        final String regexPattern = "^((\\+7|7|8)+([0-9]){10})$";
+        return Pattern.compile(regexPattern).matcher(phone).matches();
     }
 
     /**
-     * Проверить почту, вернуть true если соответствует шаблону
+     * Проверяет, соответствует ли заданный email адрес заданному шаблону.
      *
-     * @param email
-     * @return
+     * @param email email адрес, который нужно проверить.
+     * @return true, если email адрес соответствует шаблону, и false в противном случае.
      */
     private boolean isEmailValid(String email) {
-        //Добавить проверку почты
-        return true;
-
+        final String regexPattern = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$";
+        return Pattern.compile(regexPattern).matcher(email).matches();
     }
+
+    /**
+     * Метод, который связывает животное с пользователем по их идентификаторам.
+     *
+     * @param animalId идентификатор животного, которую нужно связать с пользователем
+     * @param userId   идентификатор пользователя, который будет связан с животным
+     * @throws EntityNotFoundException если животное или пользователь не найдены в базе данных
+     */
+    public void joinAnimalAndUser(long animalId, long userId) {
+        log.info("Joining animal and user with animal id: " + animalId + " and user id: " + userId);
+        Animal animal = animalRepository.findById(animalId)
+                .orElseThrow(() -> new EntityNotFoundException("Animal not found"));
+        User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("User not found"));
+        animal.setUser(user);
+        animalRepository.save(animal);
+    }
+
 }
+
+
